python/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantumliquidity"
version = "1.0.0"
description = "QuantumLiquidity Professional Trading Platform"
authors = [
    {name = "QuantumLiquidity Team"}
]
requires-python = ">=3.11"
dependencies = [
    "asyncio",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
    "ipython>=8.15.0",
    "jupyter>=1.0.0",
]

ai = [
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["quantumliquidity*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
